app/www/backoffice/src/test/javascript/spec/app/entities/xocdia-jackpot-record/xocdia-jackpot-record.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { XocdiaJackpotRecordService } from 'app/entities/xocdia-jackpot-record/xocdia-jackpot-record.service';
import { IXocdiaJackpotRecord, XocdiaJackpotRecord } from 'app/shared/model/xocdia-jackpot-record.model';

describe('Service Tests', () => {
  describe('XocdiaJackpotRecord Service', () => {
    let injector: TestBed;
    let service: XocdiaJackpotRecordService;
    let httpMock: HttpTestingController;
    let elemDefault: IXocdiaJackpotRecord;
    let expectedResult: IXocdiaJackpotRecord | IXocdiaJackpotRecord[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(XocdiaJackpotRecordService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new XocdiaJackpotRecord(0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a XocdiaJackpotRecord', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new XocdiaJackpotRecord()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a XocdiaJackpotRecord', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of XocdiaJackpotRecord', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a XocdiaJackpotRecord', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
